DinhNLHE140992_User_Authorization

+++ src/java/controller/CartController.java
@@ -46,20 +46,24 @@ public class CartController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        try (PrintWriter out = response.getWriter()) {
-
-            HttpSession session = request.getSession();
-            User u = (User) session.getAttribute("user");
-            System.out.println(u.getUserId());
-            if (u != null) {
-                CartDAO addToCartDAO = new CartDAO();
-                ArrayList<CartDetail> cartDetails = addToCartDAO.getCartByUser(u.getUserId());
-                request.setAttribute("cartDetails", cartDetails);
-                System.out.println("cartDetails" + cartDetails.size());
-                request.getRequestDispatcher("cart.jsp").forward(request, response);
-            } else {
-                response.sendRedirect("signin");
+        HttpSession session = request.getSession();
+        User u = (User) session.getAttribute("user");
+        if (u != null) {
+            //Write code here
+            try (PrintWriter out = response.getWriter()) {
+                System.out.println(u.getUserId());
+                if (u != null) {
+                    CartDAO addToCartDAO = new CartDAO();
+                    ArrayList<CartDetail> cartDetails = addToCartDAO.getCartByUser(u.getUserId());
+                    request.setAttribute("cartDetails", cartDetails);
+                    System.out.println("cartDetails" + cartDetails.size());
+                    request.getRequestDispatcher("cart.jsp").forward(request, response);
+                } else {
+                    response.sendRedirect("signin");
+                }
             }
+        } else {
+            request.getRequestDispatcher("error.jsp").forward(request, response);
         }
     }
 
@@ -67,40 +71,44 @@ public class CartController extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         HttpSession session = request.getSession();
-        CartDAO addToCartDAO = new CartDAO();
         User u = (User) session.getAttribute("user");
-        ProductDAO productDAO = new ProductDAO();       
-        if (request.getParameter("id") != null) {
-            int productID = Integer.parseInt(request.getParameter("id"));
-            if (addToCartDAO.removeFromCart(productID, u.getUserId())) {
-                System.out.println("xoa thanh cong " + productID);
-                response.sendRedirect("cart");
-            }
-        } else {
-            if (request.getParameter("quantity").equals("1")) {             
-                int proID = Integer.parseInt(request.getParameter("proID"));
-                int amount = addToCartDAO.getAmountCart(proID, u.getUserId());
-                Product product = productDAO.getProductByID(proID);
-                if(product.getAmount() == amount){
-                    String checkQuantity = "1";
-                    session.setAttribute("checkQuantity", checkQuantity);
-                    response.sendRedirect("cart");
-                }else {
-                    addToCartDAO.updateCart(proID, amount, u.getUserId());
-                    response.sendRedirect("cart");
-                }               
-            } else {
-                int proID = Integer.parseInt(request.getParameter("proID"));
-                if (addToCartDAO.getAmountCart(proID, u.getUserId()) == 1) {
-                    addToCartDAO.removeFromCart(proID, u.getUserId());
-                    response.sendRedirect("cart");
-                } else {
-                    addToCartDAO.updateCart(proID, addToCartDAO.getAmountCart(proID, u.getUserId()) - 2, u.getUserId());
+        if (u != null) {
+            //Write code here
+            CartDAO addToCartDAO = new CartDAO();
+            ProductDAO productDAO = new ProductDAO();
+            if (request.getParameter("id") != null) {
+                int productID = Integer.parseInt(request.getParameter("id"));
+                if (addToCartDAO.removeFromCart(productID, u.getUserId())) {
+                    System.out.println("xoa thanh cong " + productID);
                     response.sendRedirect("cart");
                 }
+            } else {
+                if (request.getParameter("quantity").equals("1")) {
+                    int proID = Integer.parseInt(request.getParameter("proID"));
+                    int amount = addToCartDAO.getAmountCart(proID, u.getUserId());
+                    Product product = productDAO.getProductByID(proID);
+                    if (product.getAmount() == amount) {
+                        String checkQuantity = "1";
+                        session.setAttribute("checkQuantity", checkQuantity);
+                        response.sendRedirect("cart");
+                    } else {
+                        addToCartDAO.updateCart(proID, amount, u.getUserId());
+                        response.sendRedirect("cart");
+                    }
+                } else {
+                    int proID = Integer.parseInt(request.getParameter("proID"));
+                    if (addToCartDAO.getAmountCart(proID, u.getUserId()) == 1) {
+                        addToCartDAO.removeFromCart(proID, u.getUserId());
+                        response.sendRedirect("cart");
+                    } else {
+                        addToCartDAO.updateCart(proID, addToCartDAO.getAmountCart(proID, u.getUserId()) - 2, u.getUserId());
+                        response.sendRedirect("cart");
+                    }
+                }
             }
+        } else {
+            request.getRequestDispatcher("error.jsp").forward(request, response);
         }
-        
     }
 
     /**
